Core/Gossip: port some gossip updates from TC to simplify db structure (#209)
* Core/Gossip: Drop OptionType for gossip_menu_option and rename OptionIcon to OptionNpc as it also determines functionality on client side

* Core/Gossips: Drop OptionNpcFlag for gossip_menu_option

* fix trainer and dual spec

+++ src/server/game/Server/Packets/NPCPackets.cpp
@@ -36,7 +36,7 @@ WorldPacket const* WorldPackets::NPC::GossipMessage::Write()
     for (ClientGossipOptions const& options : GossipOptions)
     {
         _worldPacket << options.ClientOption;
-        _worldPacket << options.OptionNPC;
+        _worldPacket << uint8(options.OptionNPC);
         _worldPacket << options.OptionFlags;
         _worldPacket << options.OptionCost;
